Create localization_ES.lua
Please, create localization_ES.lua and upload it to a new patch. I'm sorry for deleting the previous file. This is the corrected version of the Spanish localization. 
It would be nice if you could add my username (TradOrtiz) or my actual name (Julieta Ortiz) on the main page of the mod. Thanks!